Replace implementation of nested-loop
Previous version relied on eval, and so for example this would not work:

(let ((a #(1 2 3)))
  (nested-loop (i) (array-dimensions a)
    (format t "~a~%" (aref a i)))

since (array-dimensions a) would be eval'd at macroexpansion time.

New implementation evaluates the dimensions expression after
macroexpansion, in the local scope.

Implementation works, but is a bit imperative. Probably could be made
nicer with a reduce operation.